Mover función de detalle a listado.
