+++ app/src/main/java/edu/iesam/diarybook/features/event/presentation/eventDetail/EventDetailViewModel.kt
@@ -6,16 +6,13 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import edu.iesam.diarybook.features.event.domain.Event
 import edu.iesam.diarybook.features.event.domain.GetEventByIdUseCase
-import edu.iesam.diarybook.features.event.domain.UpdateEventOldUseCase
-import edu.iesam.diarybook.features.task.presentation.taskDetail.TaskDetailViewModel.UiState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.koin.android.annotation.KoinViewModel
 
 @KoinViewModel
 class EventDetailViewModel(
-    private val getEventByIdUseCase: GetEventByIdUseCase,
-    private val updateEventOldUseCase: UpdateEventOldUseCase
+    private val getEventByIdUseCase: GetEventByIdUseCase
 ) : ViewModel() {
 
     private val _uiState = MutableLiveData<UiState>()
@@ -25,14 +22,6 @@ class EventDetailViewModel(
         viewModelScope.launch(Dispatchers.IO) {
             val event = getEventByIdUseCase(eventId)
             _uiState.postValue(UiState(event))
-        }
-    }
-
-    fun setEventOld(event: Event, old: Boolean) {
-        viewModelScope.launch(Dispatchers.IO) {
-            updateEventOldUseCase(event.id, old)
-            val updateEventOld = event.apply { this.old = old }
-            _uiState.postValue(UiState(updateEventOld))
         }
     }
